[OS-1324] List > Replace the domain service by a read view
