+++ ddd/read_view/repository/i_liste_parcours_doctoraux.py
@@ -1,41 +1,43 @@
 # ##############################################################################
 #
-#    OSIS stands for Open Student Information System. It's an application
-#    designed to manage the core business of higher education institutions,
-#    such as universities, faculties, institutes and professional schools.
-#    The core business involves the administration of students, teachers,
-#    courses, programs and so on.
+#  OSIS stands for Open Student Information System. It's an application
+#  designed to manage the core business of higher education institutions,
+#  such as universities, faculties, institutes and professional schools.
+#  The core business involves the administration of students, teachers,
+#  courses, programs and so on.
 #
-#    Copyright (C) 2015-2024 Université catholique de Louvain (http://www.uclouvain.be)
+#  Copyright (C) 2015-2025 Université catholique de Louvain (http://www.uclouvain.be)
 #
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU General Public License as published by
-#    the Free Software Foundation, either version 3 of the License, or
-#    (at your option) any later version.
+#  This program is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
 #
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
 #
-#    A copy of this license - GNU General Public License - is available
-#    at the root of the source code of this program.  If not,
-#    see http://www.gnu.org/licenses/.
+#  A copy of this license - GNU General Public License - is available
+#  at the root of the source code of this program.  If not,
+#  see http://www.gnu.org/licenses/.
 #
 # ##############################################################################
+
 from abc import abstractmethod
 from datetime import date
 from typing import List, Optional, Tuple
 
-from admission.views import PaginatedList
 from osis_common.ddd import interface
-from parcours_doctoral.ddd.dtos import ParcoursDoctoralRechercheBODTO
+from parcours_doctoral.ddd.read_view.dto.parcours_doctoral import (
+    ListeParcoursDoctoralRechercheDTO,
+)
 
 
-class IListerTousParcoursDoctoraux(interface.DomainService):
+class IListeParcoursDoctorauxRepository(interface.ReadModelRepository):
     @classmethod
     @abstractmethod
-    def filtrer(
+    def get(
         cls,
         numero: Optional[int] = None,
         noma: Optional[str] = '',
@@ -58,5 +60,5 @@ class IListerTousParcoursDoctoraux(interface.DomainService):
         page: Optional[int] = None,
         taille_page: Optional[int] = None,
         demandeur: Optional[str] = '',
-    ) -> PaginatedList[ParcoursDoctoralRechercheBODTO]:
+    ) -> ListeParcoursDoctoralRechercheDTO:
         raise NotImplementedError
