+++ infrastructure/parcours_doctoral/read_view/handlers.py
@@ -6,7 +6,7 @@
 #    The core business involves the administration of students, teachers,
 #    courses, programs and so on.
 #
-#    Copyright (C) 2015-2024 Université catholique de Louvain (http://www.uclouvain.be)
+#    Copyright (C) 2015-2025 Université catholique de Louvain (http://www.uclouvain.be)
 #
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU General Public License as published by
@@ -23,26 +23,15 @@
 #    see http://www.gnu.org/licenses/.
 #
 # ##############################################################################
-
-from django.test import TestCase
-
-from parcours_doctoral.ddd.commands import ListerTousParcoursDoctorauxQuery
-from parcours_doctoral.infrastructure.message_bus_in_memory import (
-    message_bus_in_memory_instance,
-)
-from parcours_doctoral.infrastructure.parcours_doctoral.repository.in_memory.parcours_doctoral import (
-    ParcoursDoctoralInMemoryRepository,
+from parcours_doctoral.ddd.read_view.queries import ListerTousParcoursDoctorauxQuery
+from parcours_doctoral.ddd.read_view.use_case import *
+from parcours_doctoral.infrastructure.parcours_doctoral.read_view.repository.liste_parcours_doctoraux import (
+    ListeParcoursDoctorauxRepository,
 )
 
-
-class TestListerTousParcoursDoctoral(TestCase):
-    def setUp(self) -> None:
-        self.cmd = ListerTousParcoursDoctorauxQuery(matricule_doctorant='1')
-        self.message_bus = message_bus_in_memory_instance
-        ParcoursDoctoralInMemoryRepository.reset()
-
-    def test_should_rechercher_par_matricule(self):
-        propositions = self.message_bus.invoke(self.cmd)
-        self.assertEqual(len(propositions), 12)
-        for proposition in propositions:
-            self.assertEqual(proposition.matricule_doctorant, '1')
+COMMAND_HANDLERS = {
+    ListerTousParcoursDoctorauxQuery: lambda msg_bus, cmd: lister_parcours_doctoraux(
+        cmd,
+        lister_tous_parcours_doctoraux_service=ListeParcoursDoctorauxRepository(),
+    ),
+}
