+++ infrastructure/parcours_doctoral/read_view/repository/liste_parcours_doctoraux.py
@@ -1,28 +1,29 @@
 # ##############################################################################
 #
-#    OSIS stands for Open Student Information System. It's an application
-#    designed to manage the core business of higher education institutions,
-#    such as universities, faculties, institutes and professional schools.
-#    The core business involves the administration of students, teachers,
-#    courses, programs and so on.
+#  OSIS stands for Open Student Information System. It's an application
+#  designed to manage the core business of higher education institutions,
+#  such as universities, faculties, institutes and professional schools.
+#  The core business involves the administration of students, teachers,
+#  courses, programs and so on.
 #
-#    Copyright (C) 2015-2024 Université catholique de Louvain (http://www.uclouvain.be)
+#  Copyright (C) 2015-2025 Université catholique de Louvain (http://www.uclouvain.be)
 #
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU General Public License as published by
-#    the Free Software Foundation, either version 3 of the License, or
-#    (at your option) any later version.
+#  This program is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
 #
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
 #
-#    A copy of this license - GNU General Public License - is available
-#    at the root of the source code of this program.  If not,
-#    see http://www.gnu.org/licenses/.
+#  A copy of this license - GNU General Public License - is available
+#  at the root of the source code of this program.  If not,
+#  see http://www.gnu.org/licenses/.
 #
 # ##############################################################################
+
 from datetime import date
 from typing import Dict, List, Optional, Tuple
 
@@ -37,12 +38,15 @@ from parcours_doctoral.ddd.domain.model.enums import (
     ChoixEtapeParcoursDoctoral,
     ChoixStatutParcoursDoctoral,
 )
-from parcours_doctoral.ddd.domain.service.i_filtrer_tous_parcours_doctoraux import (
-    IListerTousParcoursDoctoraux,
-)
-from parcours_doctoral.ddd.dtos import ParcoursDoctoralRechercheBODTO
-from parcours_doctoral.ddd.dtos.formation import FormationRechercheBODTO
 from parcours_doctoral.ddd.formation.domain.model.enums import StatutActivite
+from parcours_doctoral.ddd.read_view.dto.formation import FormationRechercheDTO
+from parcours_doctoral.ddd.read_view.dto.parcours_doctoral import (
+    ListeParcoursDoctoralRechercheDTO,
+    ParcoursDoctoralRechercheDTO,
+)
+from parcours_doctoral.ddd.read_view.repository.i_liste_parcours_doctoraux import (
+    IListeParcoursDoctorauxRepository,
+)
 from parcours_doctoral.infrastructure.utils import (
     filter_doctorate_queryset_according_to_roles,
     get_entities_with_descendants_ids,
@@ -50,7 +54,7 @@ from parcours_doctoral.infrastructure.utils import (
 from parcours_doctoral.models import Activity, ParcoursDoctoral
 
 
-class ListerTousParcoursDoctoraux(IListerTousParcoursDoctoraux):
+class ListeParcoursDoctorauxRepository(IListeParcoursDoctorauxRepository):
     DATE_FIELD_BY_DATE_TYPE = {
         ChoixEtapeParcoursDoctoral.ADMISSION.name: 'created_at',
         ChoixEtapeParcoursDoctoral.CONFIRMATION.name: 'confirmationpaper__confirmation_date',
@@ -62,7 +66,7 @@ class ListerTousParcoursDoctoraux(IListerTousParcoursDoctoraux):
     }
 
     @classmethod
-    def filtrer(
+    def get(
         cls,
         numero: Optional[int] = None,
         noma: Optional[str] = '',
@@ -85,7 +89,7 @@ class ListerTousParcoursDoctoraux(IListerTousParcoursDoctoraux):
         page: Optional[int] = None,
         taille_page: Optional[int] = None,
         demandeur: Optional[str] = '',
-    ) -> PaginatedList[ParcoursDoctoralRechercheBODTO]:
+    ) -> ListeParcoursDoctoralRechercheDTO:
         language_is_french = get_language() == settings.LANGUAGE_CODE_FR
 
         qs = (
@@ -215,15 +219,15 @@ class ListerTousParcoursDoctoraux(IListerTousParcoursDoctoraux):
         for parcours_doctoral in qs:
             result.append(cls.load_dto_from_model(parcours_doctoral, language_is_french))
 
-        return result
+        return ListeParcoursDoctoralRechercheDTO(parcours_doctoraux=result)
 
     @classmethod
     def load_dto_from_model(
         cls,
         parcours_doctoral: ParcoursDoctoral,
         language_is_french: bool,
-    ) -> ParcoursDoctoralRechercheBODTO:
-        return ParcoursDoctoralRechercheBODTO(
+    ) -> ParcoursDoctoralRechercheDTO:
+        return ParcoursDoctoralRechercheDTO(
             uuid=parcours_doctoral.uuid,
             statut=parcours_doctoral.status,
             reference=parcours_doctoral.formatted_reference,  # From annotation
@@ -231,7 +235,7 @@ class ListerTousParcoursDoctoraux(IListerTousParcoursDoctoraux):
             genre_doctorant=parcours_doctoral.student.gender,
             nom_doctorant=parcours_doctoral.student.last_name,
             prenom_doctorant=parcours_doctoral.student.first_name,
-            formation=FormationRechercheBODTO(
+            formation=FormationRechercheDTO(
                 sigle=parcours_doctoral.training.acronym,
                 code=parcours_doctoral.training.partial_acronym,
                 annee=parcours_doctoral.training.academic_year.year,
